надеюсь это конец

+++ Graph/mygraph.h
@@ -125,7 +125,7 @@ Error graph_delete_vertex(Graph * graph, char * name);
 size_t _index(Vertex * vs, Vertex * v);
 Vertex ** BFS(Graph * graph, Vertex * start_v, char * name, size_t * depth);
 Vertex ** graph_BFS(Graph * graph, Vertex * start_vertex, char * name);
-int * BFA(Graph * graph, Vertex * start_vertex);
+double * BFA(Graph * graph, Vertex * start_vertex);
 Vertex ** DFS(Graph * graph, Vertex * start_v);
 Vertex ** top_sort(Graph * graph, Vertex * start_v);
